backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
sqlmodel>=0.0.22
psycopg[binary]>=3.2.0
pillow>=10.0.0
python-multipart>=0.0.9
pydantic-settings>=2.3.0
openai>=1.54.0
pymupdf>=1.24.0

boto3>=1.35.0
